Organise suitable classes into packages
