+++ src/main/java/diheng/Storage.java
@@ -9,6 +9,7 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 
+import diheng.exceptions.DiHengException;
 import diheng.tasks.Deadline;
 import diheng.tasks.Event;
 import diheng.tasks.Task;
