+++ src/main/java/diheng/exceptions/DiHengException.java
@@ -1,4 +1,4 @@
-package diheng;
+package diheng.exceptions;
 
 public class DiHengException extends Exception {
